+++ app/controllers/profiles_controller.rb
@@ -1,0 +1,24 @@
+class ProfilesController < ApplicationController
+    before_action :set_user, only: %i[show edit update]
+
+    def show
+    end
+
+    def edit
+    end
+
+    def update
+      if @user.update(user_params)
+        redirect_to profile_path, success: "ユーザーを更新しました"
+      else
+          flash.now[:danger] = "ユーザーを更新できませんでした"
+          render :edit
+      end
+    end
+
+    private
+
+    def set_user
+      @user = User.find(current_user.id)
+    end
+end
