+++ app/views/posts/show.html.erb
@@ -4,7 +4,7 @@
     <span class="badge badge-xs badge-warning">Most Popular</span><!--ラベル-->
 
     <div>
-            <iframe class="skeleton h-150 w-full"  src="https://www.youtube.com/embed/<%= find_youtube_url(@post.youtube_video) %>" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
+        <iframe class="skeleton h-150 w-full"  src="https://www.youtube.com/embed/<%= find_youtube_url(@post.youtube_video) %>" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
     </div>
 
     <div class="flex mt-5">
