+++ app/views/profiles/show.html.erb
@@ -1,0 +1,14 @@
+<div class="flex justify-center mt-8">
+  <div class="card bg-base-100 w-96 shadow-sm">
+    <figure class="px-10 pt-10 mt-5">
+      <%= image_tag 'default_profile_image.png', class: "rounded-xl" %>
+    </figure>
+    <div class="card-body items-center text-center">
+      <h2 class="card-title"><%= current_user.name %></h2>
+      <p><%= current_user.email %></p>
+      <div class="card-actions">
+        <button id="profile-edit-button" class="btn btn-primary"><%= link_to '編集', edit_profile_path %></button>
+      </div>
+    </div>
+  </div>
+</div>
