+++ app/views/shared/_header.html.erb
@@ -9,7 +9,7 @@
                 <ul
                     tabindex="0"
                     class="menu menu-sm dropdown-content bg-base-100 rounded-box z-1 mt-3 w-52 p-2 secondary text-base-content">
-                    <li><%= link_to "プロフィール", "/users/1/edit" %></li> <!--edit_user_path-->
+                    <li><%= link_to "プロフィール", profile_path %></li> <!--edit_user_path-->
                     <li><%= link_to "ブックマーク", "" %></li>
                     <div id="nav">
                     <% if current_user %>
