--- app/views/users/edit.html.erb
@@ -1,13 +1,0 @@
-<%= form_with model: @user, url: user_path, local: true do |f| %>
-   <%= f.label :email %>
-   <%= f.email_field :email, class: 'form-control mb-3' %>
-
-   <%= f.label :name %>
-   <%= f.text_field :name, class: 'form-control mb-3' %>
-
-   <%= f.label :avatar %>
-   <%= f.file_field :avatar,class: 'form-control mb-3', accept: 'image/*' %>
-   <%= f.hidden_field :avatar_cache %>
-
-   <%= f.submit "更新", class: "btn btn-primary"%>
-<%end%>
